Here is the latest v2.2 29.04.2024

+++ src/app/app.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { Model, TodoItem } from './model';
-import { throwError } from 'rxjs';
+
 
 @Component({
   selector: 'app-root',
@@ -8,32 +8,41 @@ import { throwError } from 'rxjs';
   styleUrls: ['./app.component.scss']
 })
 export class AppComponent {
+  todoItemValue:string="";
   isDisplay: boolean = true;
   user = 'Ahmet';
   model = new Model();
-
   getName() {
     return this.model.user;
   }
 
   getItems() {
-    if (!this.isDisplay) {
-      return this.model.items.filter((item) => !item.action);
+    if (this.isDisplay) {
+      return this.model.items;
     }
-    return this.model.items;
+    return this.model.items.filter((item) => !item.action);
+
   }
 
   addItem(value: string) {
-    if (value != null) {
+    if (value && value.trim() !== "") {
       this.model.items.push(new TodoItem(value, false));
-     return this.model.saveItems(); // Save items to local storage
+      this.todoItemValue = "";
+      this.model.saveItems();
     } else {
-      return throwError;
+      alert("Value cannot be blank.");
+
+  }
+  }
+
+  deleteItem(index: number) {
+    if (confirm("Are you sure you want to delete this item?")) {
+      this.model.items.splice(index, 1);
+
+      localStorage.setItem('todoItems', JSON.stringify(this.model.items));
+    } else {
+
     }
 
   }
-  deleteItem(index: number) {
-    this.model.items.splice(index, 1); // İndexe göre öğeyi listeden kaldır
-    localStorage.setItem('todoItems', JSON.stringify(this.model.items)); // Değişikliği local storage'a kaydet
-}
 }
